Infrastructure for jump-to-jump elimination in VOPs and instructions
sb!assem:*index-forwarding-label-table* maps conses of segment and of
indices in that segment to the label they jump to, if any. Its default
value, NIL, disables forwarding computation.

That table is populated with the macro (sb!assem:forward-here-to label)
with a VOP, or with (sb!assem:%forward-here-to segment label) from
instruction emitters.  Although the assembler shifts bytes and fixups
around, *indices* are preserved, and this table is updated when segments
are appended.

(sb!assem:label-final-destination label segment &optional (really t))
computes the final label the LABEL argument (which must be in SEGMENT)
jumps to at the end of any chain of unconditional jumps, unless REALLY
is NIL (in which case the initial label is returned).

sb!assem:label-position similarly gains a third optional argument to
determine whether label-final-destination should be used to straighten
jump chains out.  The default value is NIL for backward compatibility;
otherwise, it must be the segment label is in.

Code generation simply has to bind sb!assem:*index-forwarding-label-table*
to an EQUAL hash table around all assembly and segment appending and
finalization.